xmonad: fix docking issue
When undocking the laptop when lid was closed, it was suspended before
the dock even was processed and lcd was never turned on. Fixed this by
disabling suspend on lid close (not tracked in git).

Script was adjusted to always use either lcd (undocked) or just external
screen (docked). This seems to work even when the system is suspended
before a (un)dock event happens.